better handling of single quote and nifty-rific fade in fade out
Was an issue where if there was only one symbol to track it would not
render.  This was because of the way the data was returned from yahoo
when using select in query an array was not returned just a single
object so using length to loop through was not very functional.  need a
better way to handle what I did but at least it works.